rename batch conversion file
