+++ packages/payment-processor/src/index.ts
@@ -9,6 +9,7 @@ export * from './payment/near-input-data';
 export * from './payment/eth-proxy';
 export * from './payment/eth-fee-proxy';
 export * from './payment/batch-proxy';
+export * from './payment/batch-conversion-proxy';
 export * from './payment/swap-conversion-erc20';
 export * from './payment/swap-any-to-erc20';
 export * from './payment/swap-erc20';
